requirements.txt
face_recognition
Pillow
opencv-python
numpy
Flask
Flask-CORS